+++ CHANGELOG.rst
@@ -29,6 +29,8 @@ Fixed
   <https://github.com/omni-us/jsonargparse/pull/771>`__).
 - Misleading error message when a namespace is used in a list comprehension
   (`#772 <https://github.com/omni-us/jsonargparse/pull/772>`__).
+- ``omegaconf+`` parser mode failing when there are ``inf``, ``-inf`` or ``nan``
+  values (`#773 <https://github.com/omni-us/jsonargparse/pull/773>`__).
 
 
 v4.41.0 (2025-09-04)